Remove redudunt changes. Code clean up.
